moves deleteRecipe to recipeReducer, filters recipeList after delete
