+++ store/recipes/recipeReducer.js
@@ -3,11 +3,15 @@ import {
     FETCH_RECIPES_SUCCESS,
     FETCH_RECIPES_FAILURE,
     ADD_RECIPE,
+    DELETE_RECIPE_START,
+    DELETE_RECIPE_SUCCESS,
+    DELETE_RECIPE_FAILURE,
 } from "./recipeActions";
 
 const initState = {
     recipeList: [],
     isLoading: true,
+    isSubmitting: false,
     error: null,
 };
 
@@ -35,6 +39,31 @@ export const recipeReducer = (state = initState, action) => {
             return {
                 ...state,
                 recipeList: [...state.recipeList, action.payload],
+            };
+        case DELETE_RECIPE_START:
+            return {
+                ...state,
+                isSubmitting: true,
+                isLoading: true,
+                error: null,
+            };
+
+        case DELETE_RECIPE_SUCCESS:
+            return {
+                ...state,
+                isSubmitting: false,
+                isLoading: false,
+                recipeList: state.recipeList.filter(
+                    recipe => recipe.id !== action.payload,
+                ),
+            };
+
+        case DELETE_RECIPE_FAILURE:
+            return {
+                ...state,
+                isSubmitting: false,
+                isLoading: false,
+                error: action.payload,
             };
         default:
             return state;
