+++ store/singleRecipe/singleRecipeActions.js
@@ -235,28 +235,6 @@ export const fetchVersionByRevisionId = (id, revisionId) => {
     };
 };
 
-export const DELETE_RECIPE_START = "DELETE_RECIPE_START";
-export const DELETE_RECIPE_SUCCESS = "DELETE_RECIPE_SUCCESS";
-export const DELETE_RECIPE_FAILURE = "DELETE_RECIPE_FAILURE";
-
-export const deleteRecipe = id => {
-    return async dispatch => {
-        console.log("this is the id Im passing in", id);
-        dispatch({ type: DELETE_RECIPE_START });
-        try {
-            const axiosCustom = await axiosWithAuth();
-            const res = await axiosCustom.delete(`recipes/${id}`);
-            console.log("response from deleting", res.data);
-
-            dispatch({ type: DELETE_RECIPE_SUCCESS, payload: id });
-        } catch (error) {
-            console.log(error);
-            dispatch({ type: DELETE_RECIPE_FAILURE, payload: err });
-            throw error;
-        }
-    };
-};
-
 export const RESET_ALERTS = "RESET_ALERTS";
 
 export const resetAlerts = () => {
